decimal support for size and price

+++ src/main/java/com/mehmet/brokagefirm/dto/OrderDTO.java
@@ -3,12 +3,14 @@ package com.mehmet.brokagefirm.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class OrderDTO {
     private Long customerId;
     private String assetName;
     private String orderSide;
-    private Long size;
-    private Long price;
+    private BigDecimal size;
+    private BigDecimal price;
 }
